Added Auto resize on popupmenu + moved ResizableTextarea to common
