+++ src/components/common/formFields/ResizableTextArea.tsx
@@ -1,39 +1,82 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect, useRef } from 'react';
+import { useThrottledEffect } from "../helpers/Helpers"
 
-export function ResizableTextArea({ className, defaultRows, value, onChange, onBlur, placeholder }) {
-    const [rows, setRows] = useState(defaultRows)
-    const handleChange = (event) => {
-        reInitRows(event)
-        onChange(event)
-    }
-    const reInitRows = (event) => {
-        const textareaLineHeight = 28
-        const previousRows = event.target.rows
-        event.target.rows = defaultRows
-        const currentRows = ~~(event.target.scrollHeight / textareaLineHeight)
-        if (currentRows === previousRows) {
-            event.target.rows = currentRows
-        }
-        if (currentRows >= 4) {
-            event.target.rows = 4
-            event.target.scrollTop = event.target.scrollHeight
-        }
-        setRows(currentRows)
-    }
-    const handleOnBlur = (event) => {
-        setRows(defaultRows)
-        onBlur(event)
-    }
-    return (
-        <textarea
-            rows={rows}
-            value={value}
-            placeholder={placeholder}
-            className={`${className || ''} mh-30 lh-20`}
-            style={{ resize: 'none' }}
-            onChange={handleChange}
-            onBlur={handleOnBlur}
-            onFocus={reInitRows}
-        />
-    )
+interface ResizableTextareaProps {
+  minHeight?: number
+  maxHeight?: number
+  value?: string
+  onChange?: (e) => void
+  onBlur?: (e) => void
+  className?: string
+  placeholder?: string
+  lineHeight?: number
+  padding?: number
+  disabled?: boolean
+  name?: string
+  tabIndex?: number
+}
+
+export const ResizableTextarea: React.FC<ResizableTextareaProps> = ({
+  minHeight,
+  maxHeight,
+  value,
+  onChange = null,
+  onBlur = null,
+  className = '',
+  placeholder = 'Enter your text here..',
+  lineHeight = 14,
+  padding = 12,
+  disabled = false,
+  tabIndex = null,
+  ...props
+}) => {
+  const [text, setText] = useState('')
+  const _textRef = useRef(null)
+
+  useEffect(() => {
+      setText(value)
+  }, [value])
+
+  function handleChange(e) {
+      e.persist()
+      setText(e.target.value)
+      if (typeof onChange === 'function') onChange(e)
+  }
+
+  function handleBlur(e) {
+      e.persist()
+      if (typeof onBlur === 'function') onBlur(e)
+  }
+
+  useThrottledEffect(
+      () => {
+          _textRef.current.style.height = 'auto'
+          let nextHeight = _textRef.current.scrollHeight
+          if (minHeight && nextHeight < minHeight) {
+              nextHeight = minHeight
+          }
+          if (maxHeight && nextHeight > maxHeight) {
+              nextHeight = maxHeight
+          }
+          _textRef.current.style.height = nextHeight + 2 + 'px'
+      },
+      500,
+      [text],
+  )
+
+  return (
+      <textarea
+          ref={(el) => (_textRef.current = el)}
+          value={text}
+          placeholder={placeholder}
+          className={`dc__resizable-textarea ${className}`}
+          onChange={handleChange}
+          onBlur={handleBlur}
+          style={{ lineHeight: `${lineHeight}px`, padding: `${padding}px` }}
+          spellCheck={false}
+          disabled={disabled}
+          tabIndex={tabIndex}
+          {...props}
+      />
+  )
 }
