+++ src/components/common/index.ts
@@ -51,3 +51,4 @@ export * from './helpers/compareVersion'
 export * from './header/PageHeader'
 export * from './error/ErrorBar'
 export * from './DetectBottom'
+export * from './formFields/ResizableTextArea'
